refactor: 캐시 추상화 및 의존성 최적화
