+++ src/main/kotlin/com/deepromeet/atcha/transit/infrastructure/cache/LastRouteRedisCache.kt
@@ -1,14 +1,16 @@
 package com.deepromeet.atcha.transit.infrastructure.cache
 
 import com.deepromeet.atcha.transit.api.response.LastRoutesResponse
+import com.deepromeet.atcha.transit.domain.LastRouteCache
 import org.springframework.data.redis.core.RedisTemplate
 import org.springframework.data.redis.core.ScanOptions
 import org.springframework.stereotype.Component
+import java.time.Duration
 
 @Component
 class LastRouteRedisCache(
     private val lastRouteRedisTemplate: RedisTemplate<String, LastRoutesResponse>
-) {
+) : LastRouteCache {
     fun processRoutes(action: (LastRoutesResponse) -> Unit) {
         val valueOps = lastRouteRedisTemplate.opsForValue()
 
@@ -33,8 +35,12 @@ class LastRouteRedisCache(
         }
     }
 
-    fun save(lastRoutesResponse: LastRoutesResponse) {
-        lastRouteRedisTemplate.opsForValue().set(getKey(lastRoutesResponse.routeId), lastRoutesResponse)
+    override fun get(routeId: String): LastRoutesResponse? {
+        return lastRouteRedisTemplate.opsForValue().get(getKey(routeId))
+    }
+
+    override fun cache(route: LastRoutesResponse) {
+        lastRouteRedisTemplate.opsForValue().set(getKey(route.routeId), route, Duration.ofHours(12))
     }
 
     fun getKey(routeId: String): String {
